+++ test/model.mapreduce.test.js
@@ -174,8 +174,7 @@ describe('model: mapreduce:', function() {
       verbose: false
     };
 
-    const res = await MR.mapReduce(o);
-    console.log(res);
+    assert.deepEqual(await MR.mapReduce(o), []);
     assert.equal(typeof stats, 'undefined');
   });